+++ wazo_webhookd/plugins/config/http.py
@@ -3,6 +3,7 @@
 
 from flask import request
 from wazo_webhookd.rest_api import AuthResource
+from wazo_webhookd.auth import required_master_tenant
 from xivo.auth_verifier import required_acl
 from jsonpatch import JsonPatch
 
@@ -13,10 +14,12 @@ class ConfigResource(AuthResource):
     def __init__(self, config_service):
         self._config_service = config_service
 
+    @required_master_tenant()
     @required_acl('webhookd.config.read')
     def get(self):
         return self._config_service.get_config(), 200
 
+    @required_master_tenant()
     @required_acl('webhookd.config.update')
     def patch(self):
         config_patch = config_patch_schema.load(request.get_json(), many=True)